Refactored version, processes now communicate to sync actions

+++ transmitter.c
@@ -12,47 +12,62 @@
 #include <time.h>
 
 #define DEBUG 0
+#define SLP 0
 
 #define DBG(expr) if (DEBUG) {expr}
-
-#define NANO_SLEEP 130000
+#define SLEEP if (SLP) {sleep(1);}
 
 
 
 void to_bits(char* buf, int buf_size, int pid) {
-    time_t begin;
-    time(&begin);
+
+    time_t start;
+    time(&start);
+
+    int res = -1;
+    sigset_t waitset;
+    siginfo_t siginfo;
+
+    sigemptyset(&waitset);
+    sigaddset(&waitset, SIGUSR1);
+    sigaddset(&waitset, SIGUSR2);
+    res = sigprocmask(SIG_BLOCK, &waitset, NULL);
+    if (res == -1) {
+        fprintf(stderr, "Cannot block signals\n");
+        exit(-1);
+    }
 
     for (int i = 0; i < buf_size; ++i) {
         char byte = buf[i];
         for (unsigned j = 0; j < sizeof(char) * 8; ++j) {
+
             unsigned mask = 1 << (sizeof(char) * 8 - 1 - j);
             char bit = (mask & byte) >> (sizeof(char) * 8 - 1 - j);
-            DBG(fprintf(stderr, "%d", bit);)
-            struct timespec time;
-            time.tv_nsec = NANO_SLEEP;
-            time.tv_sec = 0;
-            struct timespec rem;
-            nanosleep(&time, &rem);
-            //sleep(1);
+
+            SLEEP
             if (bit == 0) {
-                kill(pid, SIGUSR1);
-            } else kill(pid, SIGUSR2);
+                res = kill(pid, SIGUSR1);
+                assert(res != -1);
+                DBG(fprintf(stderr, "Sending signal 0!\n");)
+            } else {
+                res = kill(pid, SIGUSR2);
+                assert(res != -1);
+                DBG(fprintf(stderr, "Sending signal 1!\n");)
+            }
+             res = sigwaitinfo(&waitset, &siginfo);
+             assert(res != -1);
+             DBG(fprintf(stderr, "Received responce!\n");)
         }
-        DBG(fprintf(stderr, " ");)
     }
+
+    DBG(fprintf(stderr, "\n");)
     kill(pid, SIGTERM);
+
     time_t end;
     time(&end);
 
-    //fprintf(stderr, "buf_size %d\n", buf_size);
-    unsigned speed = buf_size / difftime(end, begin);
-
-    printf("Speed: %u Bytes/sec\n", speed);
-    // clock_t end = clock();
-    // double time_spent = (double)(end - begin) / CLOCKS_PER_SEC;
-    // unsigned speed = buf_size / time_spent;
-    // printf("Speed: %u Bytes/sec\n", speed);
+    unsigned speed = buf_size / difftime(end, start);
+    fprintf(stdout, "Speed: %u Bytes/sec\n", speed);
 }
 
 // void send_size_queued(int size, int pid) {
@@ -71,26 +86,44 @@ void to_bits(char* buf, int buf_size, int pid) {
 // }
 
 void send_size(int input_size, int pid) {
+
+    int res = -1;
+    sigset_t waitset;
+    siginfo_t siginfo;
+
+    sigemptyset(&waitset);
+    sigaddset(&waitset, SIGUSR1);
+    sigaddset(&waitset, SIGUSR2);
+    res = sigprocmask(SIG_BLOCK, &waitset, NULL);
+    if (res == -1) {
+        fprintf(stderr, "Cannot block signals\n");
+        exit(-1);
+    }
+    
     for (unsigned j = 0; j < sizeof(int) * 8; ++j) {
-            unsigned mask = 1 << (sizeof(int) * 8 - 1 - j);
-            char bit = (mask & input_size) >> (sizeof(int) * 8 - 1 - j);
-            DBG(fprintf(stderr, "%d", bit);)
-            struct timespec time;
-            time.tv_nsec = NANO_SLEEP;
-            time.tv_sec = 0;
-            struct timespec rem;
-            nanosleep(&time, &rem);
-            //sleep(1);
-            if (bit == 0) {
-                kill(pid, SIGUSR1);
-                //fprintf(stderr, "Sending signal 0!\n");
-            } else {
-                kill(pid, SIGUSR2);
-                //fprintf(stderr, "Sending signal 1!\n");
-            }
+
+        unsigned mask = 1 << (sizeof(int) * 8 - 1 - j);
+        char bit = (mask & input_size) >> (sizeof(int) * 8 - 1 - j);
+        DBG(fprintf(stderr, "%d", bit);)
+        
+        SLEEP
+        if (bit == 0) {
+            res = kill(pid, SIGUSR1);
+            assert(res != -1);
+            DBG(fprintf(stderr, "Sending signal 0!\n");)
+        } else {
+            res = kill(pid, SIGUSR2);
+            assert(res != -1);
+            DBG(fprintf(stderr, "Sending signal 1!\n");)
         }
-        DBG(fprintf(stderr, "\n");)
+            res = sigwaitinfo(&waitset, &siginfo);
+            assert(res != -1);
+            DBG(fprintf(stderr, "Received responce!\n");)
+    }
+    DBG(fprintf(stderr, "\n");)
 }
+
+
 int main (int argc, char** argv) {
 
     assert(argc == 3);
@@ -101,9 +134,11 @@ int main (int argc, char** argv) {
 
     int fd = open(file, O_RDONLY);
     assert(fd > 0);
+
     struct stat input;
     fstat(fd, &input);
     int input_size = input.st_size;
+    assert(input_size > 0);
 
     char* buf = (char*) calloc(input_size, sizeof(char));
     assert(buf != NULL);
@@ -112,11 +147,7 @@ int main (int argc, char** argv) {
     assert(n_read == input_size);
     buf[input_size] = '\0';
 
-    //fprintf(stderr, "Buf read: %s\n", buf);
-
-    //send_size_queued(input_size, pid);
     send_size(input_size, pid);
-    
     to_bits(buf, input_size, pid);
 
     close(fd);
